F1c(1): D9 and D16, u16 group indices and dynamic capture slots
One logical commit (review: 17-255 captures must never be silently wrong
between commits), two parts.

(a) Group indices are u16: ast.group_index, the lexer's backref token,
parser.group_counter (error.TooManyCaptures past 65535 instead of
wrapping, which fixes D16: 256 sequential groups panicked in ReleaseSafe
and wrapped to capture 0 in ReleaseFast), GroupNameEntry/NamedGroup
index, generator/compiler group_count, regex.zig, and the bytecode:
SAVE_START, SAVE_END, CLEAR_CAPTURE, BACK_REF(_I) carry a u16 group (size
3), SAVE_*_NAMED u16 + u32 (size 7). C API: group_index parameters and
zregex_named_group_index are usize; zregex.mjs signatures updated.

(b) Capture slots are sized to the pattern: RecursiveMatcher keeps
group_count + 1 slots (inline up to 16, heap beyond), MatchResult is just
{matched, end_pos}, the captures are read from the matcher after a
successful match (captureSlice), and lookaround snapshots live on a heap
stack instead of in each recursion frame. Matcher gets the group count
from CompileResult (initWithGroups). Parser: branch_stack is an ArrayList
and GroupNameEntry.branch_path an owned slice; MAX_ALTERNATION_DEPTH and
AlternationTooDeep are gone, nesting_depth is the effective limit, and
"MAX_NESTING_DEPTH levels fit in 1 MiB (release builds)" runs (Release;
Debug still skips until F2).

Sites found by the pre-edit grep of MAX_CAPTURE_GROUPS and the literal
16, all changed:
- recursive_matcher.zig:22 (MAX_CAPTURE_GROUPS = 16), :61/:65
  (MatchResult.captures array and getCapture bound), :96/:121 (matcher
  captures array and init), :415/:429/:446 (SAVE_START/SAVE_END/
  CLEAR_CAPTURE bounds), :1506 (checkBackRef bound);
- matcher.zig:144/:145 (alloc(Capture, 16) and its copy loop).
Also found outside that grep and changed: c_api.zig zregex_match_group
rejected group_index >= 10; zregex.mjs skipped groups past 255.

Fixed on the way (pre-existing, found by differential-v8): a
backreference to a group re-entered but not closed yet (new start, old
end) overflowed `end - start` in checkBackRef, a panic in safe builds;
it now matches empty, as the spec's cleared capture does. It was all 61
crashes of the differential reference.

Tests: tests/captures_tests.zig (17 and 200 sequential groups and 200
nested with exact captures; D16 at 255/256/257/300 groups checking the
last group's indices; \17, \200, \300; 65535 groups compile, 65536 is
TooManyCaptures; lookaround snapshots past the inline slots; allocation
failures with heap slots; the backreference overflow), and in
recursive_matcher.zig 1000 groups with exact captures and \1000 (limit
lifted, 64 MiB thread: through Regex, 1000 groups hit the recursion
limit, a defined error, D14/F6a). Tests updated for the new instruction
sizes; parser tests now deinit the parser. The D14 regression test on a
1 MiB stack is enabled (it passes now, see below). Fuzz stress: a
many-groups generator (1-300 groups, some with a backreference to the
last one).

Measured (probes, not committed): stack per nested capturing group (T15)
~3.4 KiB in ReleaseSafe (was ~49) and ~21.7 KiB in Debug (was ~147);
D15 ~1.8 KiB per level (was ~25.3) and ~11.2 KiB in Debug (was ~75.5),
so ()\1{1000} returns RecursionLimitExceeded on 8 MiB in ReleaseSafe;
the D14 body pattern needs ~103 KiB (ReleaseSafe) / ~615 KiB (Debug).
Bench: about 3x faster on every throughput case, adversarial 28-35 ms
(were 44-48). differential-v8: no new divergence on any previously
compared case; 0 crashes (were 61).

test262: T15 and T16 (4 entries) pass, 0 regressions.

+++ src/codegen/compiler.zig
@@ -32,7 +32,7 @@ pub const CompileResult = struct {
     /// distinguish "group N doesn't exist in this pattern" from "group N
     /// exists but didn't participate in this match" -- e.g. for `$N`
     /// substitution in `Regex.replace`/`replaceAll`.
-    group_count: u8,
+    group_count: u16,
     allocator: Allocator,
 
     /// Free the compilation result
